requirements.txt
PyYAML==5.4.1
requests==2.26.0
aiohttp==3.10.9